requirements.txt
streamlit
streamlit-chat
torch
torchvision
Pillow
numpy
pyyaml
openai
python-dotenv